+++ app/src/androidTest/java/com/example/moneymanager/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.expensemanagerapp;
+package com.example.moneymanager;
 
 import android.content.Context;
