+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Expense Manager App</string>
+    <string name="app_name">Money Manager</string>
 </resources>
