+++ build.gradle.kts
@@ -2,3 +2,4 @@
 plugins {
     id("com.android.application") version "8.2.0" apply false
 }
+
